# src/main/resources/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bus Booking System API",
    "description": "Єдина документація для системи бронювання квитків. Включає авторизацію та роботу з містами.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Локальний сервер"
    }
  ],
...
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Authorization"],
        "summary": "Реєстрація",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } }
          }
        },
        "responses": { "200": { "description": "Успіх" } }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authorization"],
        "summary": "Логін",
        "description": "Працює захист від Brute Force (Caffeine Cache).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } }
          }
        },
        "responses": {
          "200": { "description": "Повертає JWT токен" },
          "429": { "description": "Забагато спроб (IP заблоковано)" }
        }
      }
    },
...
    "/api/auth/password": {
      "put": {
        "tags": ["Authorization"],
        "summary": "Зміна пароля",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ChangePasswordRequest" } }
          }
        },
        "responses": { "200": { "description": "Пароль оновлено" } }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "password": { "type": "string" }
        }
      },
...
      "ChangePasswordRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "oldPassword": { "type": "string" },
          "newPassword": { "type": "string" }
        }
      },
...
    }
  }
}